fix(relatorios): aplicar review do PDF clinico
