+++ .memory/rules.md
@@ -84,9 +84,9 @@ const FREQUENCIES = ['daily', 'weekly']
 **Rule:** Always follow Validate -> Record -> Decrement order for dose registration. Validate stock first, then create log, then decrease stock.
 **Source:** journal/archive
 
-### R-024: Daily PDF Adherence Must Sum Quantities [HIGH]
-**Rule:** When rendering daily adherence in PDFs, sum `expectedQuantity` from taken/missed dose slots and exclude future `scheduledDoses`. Never use raw array lengths for daily dose totals when the protocol can take more than one pill per intake.
-**Source:** PR #421 PDF consultation patch (2026-03-24)
+### R-024: Daily PDF Adherence Must Count Dose Events [HIGH]
+**Rule:** In the clinical PDF daily table, count completed vs expected dose events for each day and exclude future `scheduledDoses`. Do not reuse monthly aggregate totals per row, and do not switch this table to pill-quantity math unless the UI and labels are redesigned for that unit.
+**Source:** PR #422 review follow-up (2026-03-25)
 
 ### R-025: Fallback Patient Identity From Email Handle [MEDIUM]
 **Rule:** If the profile has no display name, derive the PDF patient label from the email local-part before falling back to `"Paciente"`. This keeps consultation PDFs clinically useful even before the database has a dedicated name field.
